Axios post to API throwing error
