+++ client/src/jade/next-distro-fe/form/form.jade
@@ -1,6 +1,6 @@
 // One "tab" for each step in the form (main form tab):
 .tab
-  input(type="hidden", name=api.formUUID, value=api.formUUID)
+  input(type="hidden", name="formUUID", value=api.formUUID)
   each field in api.questions
     if field.questionType === "input"
       include partials/_input
